+++ ReleaseNotes.md
@@ -1,8 +1,12 @@
+### New in 0.0.4 (Released 2015/09/23)
+* Use FilePath and WorkingDirectory for file locations
+* Add bucket and keys to Verbose logs about the transfer details
+
 ### New in 0.0.3 (Released 2015/09/22)
-* Progress Events
+* Logging Progress-Events
 
 ### New in 0.0.2 (Released 2015/09/18)
-* Extensions methods
+* Extensions methods for settings
 
 ### New in 0.0.1 (Released 2015/09/16)
 * First release.